Removed Sass dependancies and tasks
This project depends greatly on writing css for responsiveness. This
can be achieved using sass but for the sake of this project I decided
to write everything in css to show knowledge of writing css vs using
sass and compiling it to css. I removed all the sass and css
dependancies from the package.json file and removed all the sass and
css tasks from gulp file.js except to watch the css folder for changes
so that the browse-sync will update in the development server.